requirements.txt
opencv-python
tensorflow-gpu
sklearn
numpy
